@@ -1,17 +1,23 @@
 {
   "dependencies": {
-    "@rails/webpacker": "^3.0.2",
+    "@rails/webpacker": "^3.2.0",
     "babel-preset-react": "^6.24.1",
     "coffeescript": "1.12.7",
+    "connected-react-router": "^4.3.0",
     "lodash": "^4.17.4",
+    "postcss-smart-import": "^0.7.6",
     "prop-types": "^15.6.0",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
-    "react_ujs": "^2.4.2",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "react-hot-loader": "^3.1.3",
+    "react-redux": "^5.0.6",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.2.0",
     "reqwest": "^2.0.5",
     "superagent": "^3.8.1"
   },
   "devDependencies": {
-    "webpack-dev-server": "^2.9.4"
+    "redux-logger": "^3.0.6",
+    "webpack-dev-server": "^2.9.7"
   }
 }
