@@ -1,4 +1,30 @@
-// Support component names relative to this directory:
-var componentRequireContext = require.context("components", true)
-var ReactRailsUJS = require("react_ujs")
-ReactRailsUJS.useContext(componentRequireContext)
+/* eslint no-console:0 */
+// This file is automatically compiled by Webpack, along with any other files
+// present in this directory. You're encouraged to place your actual application logic in
+// a relevant structure within app/javascript and only use these pack files to reference
+// that code so it'll be compiled.
+//
+// To reference this file, add <%= javascript_pack_tag 'application' %> to the appropriate
+// layout file, like app/views/layouts/application.html.erb
+
+import React from 'react';
+import { render } from 'react-dom';
+import { Provider } from 'react-redux';
+import { AppContainer } from 'react-hot-loader';
+
+import store from '../store/index';
+import routes from '../routes';
+
+render(
+  <Provider store={store}>{routes}</Provider>,
+  document.querySelector("#app")
+);
+
+if (module && module.hot) {
+  module.hot.accept('./containers/app', () => {
+    render(
+      <Provider store={store}>{routes}</Provider>,
+      document.querySelector('#app')
+    );
+  });
+}
