@@ -1,15 +1,13 @@
 Rails.application.routes.draw do
   root to: 'home#index'
-  get 'setup', to: 'home#setup'
 
   # omniauth callbacks
-  get '/auth/create', to: 'auth#create'
   get '/auth/:provider/callback', to: 'auth#callback'
-  post '/auth/connect', to: 'auth#connect_to_guest_user'
-  get '/complete', to: 'home#complete'
-  
-  # Imports
-  post 'imports/create', to: 'imports#create'
+  post 'auth/connect', to: 'auth#connect_to_guest_user'
+
+  namespace :api do
+    post 'imports/create', to: 'imports#create'
+  end
 
   # ActiveAdmin stuff
   devise_for :admin_users, ActiveAdmin::Devise.config
